Fix write file for new version of Evernote
It would seem that the latest version of Evernote changes theAttachment to just attachment.

I found this because running the script would throw an exception message. Digging into the Applescript dictionary I found that attachment is the proper noun to use to reference an attachment.

I also wrapped this around a try block so that if a file write fails it doesn't kill the entire script.